+++ admin/translations/_template.inc.php
@@ -425,4 +425,5 @@ return [
 	'Data table' => 'Xx',
 	'Edit form' => 'Xx',
 	'Ask %s' => 'Xx %s',
+	'Views' => 'Xx',
 ];